\name{xterm256-package}
\Rdversion{1.1}
\alias{xterm256-package}
\alias{xterm256}
\docType{package}
\title{
Support for xterm 256 escape sequences
}
\description{
Support for xterm256 escape sequences, enabling R
	functions to take advantage of foreground color, background color, 
	in capable terminals
}
\details{
\tabular{ll}{
Package: \tab xterm256\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2009-11-22\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Romain Francois

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
	\url{http://frexx.de/xterm-256-notes/}
}
\keyword{ package }

